#1213 - fix a lot of stuff
