+++ src/frontend/src/apis/finance.api.ts
@@ -135,10 +135,9 @@ export const downloadImage = async (fileId: string): Promise<File> => {
 /**
  * Reports a given dollar amount representing a new account credit
  *
- * @param id the id of the user who is being reimbursed
- * @param formData the dollar amount being reimbursed
+ * @param amount the dollar amount being reimbursed
  * @returns a reimbursement with the given user and dollar amount
  */
-export const reportRefund = (id: string, refundAmount: number) => {
-  return axios.post(apiUrls.financeReportRefund(id), refundAmount);
+export const reportRefund = (amount: number) => {
+  return axios.post(apiUrls.financeReportRefund(), { amount });
 };
